# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentShield",
    "description": "Local security reverse proxy for coding-agent traffic",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get application settings",
        "description": "Retrieve application configuration, detector limits, and provider credential presence without revealing secret keys.",
        "operationId": "get_settings_api_v1_settings_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-AgentShield-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Agentshield-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update application settings",
        "description": "Update runtime settings overrides stored in SQLite.",
        "operationId": "update_settings_api_v1_settings_put",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-AgentShield-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Agentshield-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SettingsResponse": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "proxy_max_body_bytes": {
            "type": "integer",
            "title": "Proxy Max Body Bytes"
          },
          "sse_max_event_bytes": {
            "type": "integer",
            "title": "Sse Max Event Bytes"
          },
          "detector_timeout_seconds": {
            "type": "number",
            "title": "Detector Timeout Seconds"
          },
          "pseudonym_ttl_seconds": {
            "type": "integer",
            "title": "Pseudonym Ttl Seconds"
          },
          "approval_timeout_seconds": {
            "type": "number",
            "title": "Approval Timeout Seconds"
          },
          "pii_detect_ip_addresses": {
            "type": "boolean",
            "title": "Pii Detect Ip Addresses"
          },
          "pii_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pii Languages"
          },
          "custom_terms_count": {
            "type": "integer",
            "title": "Custom Terms Count"
          },
          "openai_configured": {
            "type": "boolean",
            "title": "Openai Configured"
          },
          "anthropic_configured": {
            "type": "boolean",
            "title": "Anthropic Configured"
          },
          "protection_limits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Protection Limits"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "host",
          "port",
          "proxy_max_body_bytes",
          "sse_max_event_bytes",
          "detector_timeout_seconds",
          "pseudonym_ttl_seconds",
          "approval_timeout_seconds",
          "pii_detect_ip_addresses",
          "pii_languages",
          "custom_terms_count",
          "openai_configured",
          "anthropic_configured",
          "protection_limits"
        ],
        "title": "SettingsResponse",
        "description": "Sanitized application settings with protection boundaries and credential presence."
      },
      "SettingsUpdateRequest": {
        "properties": {
          "profile": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(audit|balanced|strict)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "approval_timeout_seconds": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Timeout Seconds"
          },
          "pseudonym_ttl_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400.0,
                "minimum": 60.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudonym Ttl Seconds"
          }
        },
        "type": "object",
        "title": "SettingsUpdateRequest",
        "description": "Payload to update runtime settings overrides."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}